refactor(companion): extract AvailabilityListScreen with AppPressable migration
